arm: Power off non-root CPUs before shutting down the hypervisor
When handing back a CPU previously used by a non-root cell directly to
Linux during hypervisor shutdown, we have to power it off when Linux is
using PSCI. Otherwise, the CPU_ON command issued by Linux later on to
gain control over the CPU again may not work as expected.

All our supported boards except for the Versatile Express come with PSCI
support. Try both v0.2 and the v0.1 encoding of U-Boot for CPU_OFF. At
least one of them must work.

The GICC is not reset by the power-down, thus we have to reset it shut
it down explicitly.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>