driver: Update CR4 on x86 after hypervisor enabling/disabling
Linux gained CR4 shadowing in 4.0: it now keeps a software copy of the
mm's CR4 state and only write updates on changes. But this also means it
will miss changes to this register when enabling or disabling Jailhouse.
That will cause sporadic access violations or false reporting of
virtualization-busy CPUs on Jailhouse enable.

Fix it by re-initializing the shadow on x86 after returning from the
hypervisor entry or the disable command. We are still protected from
interrupts at that point, so nothing can leak to other tasks.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>