x86: Fix hypercall dispatching assembly
The macro reads jailhouse_use_vmcall, a bool variable that is mapped on
a byte on x86. Specify the width to avoid that we test more than we
should.

This bug was revealed by adding further global bool variables that
happen to line up after jailhouse_use_vmcall. The current memory layout
may make the issue harmless in practice.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>