x86: Allow access to Focus Processor Checking bit in APIC SVR
The Intel manual says: "In Pentium 4 and Intel Xeon processors, this bit
is reserved and should be cleared to 0." It apparently refers to the
first Xeon series here, not newer ones that support IA32e. Linux sets
this bit on x86-64 unconditionally for more than a decade. There are no
availability restrictions mentioned for AMD at all.

So let's release this bit to the cells because it cannot cause any harm
to the system or the hypervisor.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>