arm: Remove ancient compiler bug test via __asmeq
This macro was once copied in from the Linux kernel. There it tries to
catch buggy gcc 3.x versions that didn't follow the specified register
assignments (see https://gcc.gnu.org/bugzilla/show_bug.cgi?id=15089).

This bug is now 10 years old, fixed, and affected compilers that weren't
even aware of the virt extensions for ARMv7 that we depend on anyway. So
let's remove it.

This also removes a GPL'ed line of code, thus enables a dual-licensing
of the file.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>