inmates: Define entry point in linker script
Export the reset address as symbols and define them as entry point of
our inmates in the linker scripts. We will bundle the headers together
with the other library objects in archives, and defining entry points
will ensure that the related sections will be included in the final
binary. This will simplify the inmate rules significantly.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>