x86: Flush pending events when reprogramming the VT-d error interrupt
There seems to be the risk of in-flight error events still using the
address and data registers while we reprogram them. In practice, this
shouldn't happen on a correctly configured system because all valid
interrupt sources are silenced at this point. Nevertheless, play safe,
just like Linux does.

However, there is no reason to also read back after unmasking (like
Linux does) because the hardware injects pending events when the mask is
cleared.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>