arm64: PSCI support for SMP on AArch64
On AArch64 we pretty much rely on PSCI being present for SMP
support (turning multiple cores on and off). This patch implements
the helpers needed for SMP and plugs in the PSCI code from AArch32.

On AArch64 PSCI calls can be issued via SMC64 as well, contrary to
AArch32 which uses SMC32 calls only. We add the changes necessary to
support the requests that are used by a Linux root cell. CPU hotplug now
is still working after we enable Jailhouse.

Based on patch by Antonios Motakis.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>