arm64: plug the irqchip and GICv2 code from AArch32
We plug the irqchip handling code, and the GICv2 implementation from
AArch32. arch_handle_sgi and arch_handle_phys_irq are only provided as
stubs because they will be replaced with the arm-common versions soon.

GICv3 is slightly trickier; it makes heavier use of the sysregs, so
we will need to review more carefully that the sysregs macros do
the right thing.

Based on patch by Antonios Motakis.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>