tools: config-create: Do not enter infinite over disabled PCI devices
If a PCI device is disabled, e.g. a secondary GPU, we may not see its
config space anymore while it is still listed. With config all 0xff, we
will then enter an infinite loop while trying to make sense of the
capability list. Prevent this, issuing a warning that we will skip this
device.

Reported-by: Thomas Pettinger <thomas.pettinger@tum.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>