arm: Remove return code from irqchip_cpu_reset
Only the GICv3 actually returns a non-zero value, and that only if the
reset was called before the irqchip was fully initialized, i.e. as part
of an error rollback triggered via irqchip_cpu_shutdown. Therefore,
remove the error return code and only avoid the v3's cpu_reset handler
does anything in case it was not yet initialized.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>