add recommendation for use of Fixes: tag in bug-fix commits
To improve the review of bug-fixes a standardized reference to the
bug-introducing commit, which the Fixes: tag provides, can make life
a lot simpler - and it allows some level of automation.

While strongly adviseable it is only recommended and not set to mandatory
notably as it is better *not* to reference a commit if unsure than to
reference a wrong commit.

Signed-off-by: Nicholas Mc Guire <hofrat@osadl.org>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>