x86: Require VT-d support also for QEMU
QEMU 2.7 will have all the features we need to run Jailhouse with VT-d
IOMMU. Remove the special cases from the code and document the new
command line.

The requirement on the kernel version is not exact but a safe guess
based on tests.

Support for AMD-Vi (IOMMU) emulation in QEMU is still work in progress.
So we leave the corresponding checks for that setup in place.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>