arm: Switch to new CPU management services
This patch became larger than usual ones, but I haven't found a way to
make it smaller. That's due to the fact it replaces the old, racy CPU
onlining/offlining/reset logic with the one derived from x86.

Along this comes a simplification of the smp interface we only have to
cater those few SoCs without proper PSCI support: it is no longer
required to add stubs for PSCI-based platforms. Consequently, those for
TK1 and Allwinner A20 are removed.

Now obsolete PSCI fragments will be purged in a separate patch.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>