x86: Fix cpuid.01h results regarding OSXSAVE
As we execute cpuid in the host on behalf of the guest, we have to
derive the OSXSAVE feature bit from the guest state of CR4 which
controls the availability of this while the guest is running.

Currently, if XSAVE is available and Jailhouse enables OSXSAVE for
itself in CR4, we always report it as set to the guest as well, causing
crashes or root cell applications that check for the availability via
cpuid and then run xgetbv.

Fixes: 10c6aac6eff4 ("x86: Always intercept cpuid")

Reported-by: Swaraj Dube <swaraj.dube@intel.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>