hypervisor, driver: let the driver choose cell ids
If the cell id is chosen by the driver instead of the hypervisor, the
cell id gets part of the cell configuration.

As part of the configuration, move the cell id from hypervisor's struct
cell to struct jailhouse_cell_desc, as this structure is passed from the
driver to the hypervisor.

Let the hypervisor as well as the driver check collisions of cells.
Both, cell name and cell id are still both unique values.

Also align jailhouse-cell-linux to the changed memory layout.

Signed-off-by: Ralf Ramsauer <ralf@ramses-pyramidenbau.de>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>