arm: Reject unknown SMC invocations
Reject every SMC requests except for the emulated PSCI services. This is
required because we have generally no idea about what services are
behind an interface, and even the calling convention is undefined.
Better fail loudly and then resolve the valid cases on a per-platform
basis.

This obsoletes the smc stub in exception.S, so remove it.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>