configs: Add non-root Linux demo cells and device trees for Banana Pi
Use the separate UART 7 for the inmate. Access to the CCU is granted to
allow Linux to enable the related clock gate. We do not control the
access on a bit-level, so this remains a hack. However, we can at least
avoid that Linux touches the wrong bits by describing only the relevant
one in the inmate device tree.

On ARM, we have the need to align the virtual start address of the
identity-mapped memory to 128 MB. Furthermore, Linux won't boot if that
region starts at 0xf0000000 or higher.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>