core: add divider to jailhouse_debug_console
Add an additional divider variable to the uart structure that allows to
dynamically set the uart baudrate.

If the divider is zero, the hypervisor will skip UART initialisation and
assume that the UART was already initialised by Linux.

For most scenarios, this is the default case and the reason why .divider
values are comments inside system configurations that represent common
default values, if initialisation is required.

Those default values are:
  - arm: Banana Pi  (0x0d, 8250)
  - arm: Jetson TK1 (0xdd, tegra, yet unused)
  - x86: PIO        (0x1, means 115200 (?))
  - x86: OXPCIE     (0x22, see h87i.c)

Additionally, the introduction of the divider gives us the chance to get
rid of CONFIG_SERIAL_OXPCIE952 in hypervisor code.

Signed-off-by: Ralf Ramsauer <ralf.ramsauer@oth-regensburg.de>
Cc: Jan Kiszka <jan.kiszka@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>