arm: Rename SMP hooks to platform-specific extensions
The SMP hooks so far only handle the variations of the VExpress platform
related to CPU onlining/offlining. As we are going to add more hooks for
other platforms, some SMP-unrelated, it makes sense to prefix all hooks
with "mach" instead of "smp". This is also consistent with CONFIG switch
naming.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>