configs: Enable inter-cell communication for Jetson TK1
Add ivshmem devices to root cell and linux-demo for networking purposes.
A virtual PCI host is introduced and INTx SPIs are carefully selected
for both sides so that they fall into the supported set of the GICD and
do not overlap with actually used interrupts.

We need to go the vPCI path because Linux does not allow us to inject
any virtual devices into the physical bus, see also
https://www.mail-archive.com/jailhouse-dev@googlegroups.com/msg01015.html.
To enable the second PCI host controller, multi-domain PCI support is
required for the Jetson, thus a small patch to the related Kconfig file
of the kernel.

The K1 only supports 160 SPIs, so adjust the root cell mask at this
chance as well.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>