arm64: Rework entry code to support non-identity mapping
Now that the hypervisor core is prepare for it, also convert the entry
path.

First, we have to extract some parameters from the hypervisor header and
the configuration, namely hypervisor and UART physical addresses. Their
virtual addresses are not build-time defined, both using the top 1-GB
page in the hypervisor address space. That reduces the likelihood of
address conflicts for the ID mapping of the trampoline page.

Our bootstrap page table now also has to contain an ID map of the
trampoline code. As hypervisor and UART share the same 1-GB page, they
can also share their level-1 and level-2 page tables. We map both as
2-MB pages. The trampoline page is mapped as 1-GB page and, thus,
requires an additional page table page. This structuring also simplifies
the setup because we no longer need to differentiate between various
cases.

The UART virtual address in the hypervisor header is now updated by the
entry code, taking the fixed 2-MB base address as well as the offset
into that page - based on the physical address - into account.

This change overcomes the ugly need to define a target-specific
JAILHOUSE_BASE in config.h.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>