x86: Sanitize printk format strings
This fixes all findings when enabling __attribute__((format(...))) for
printk and panic_printk. All of them were practically harmless so far,
even the missing parameter in amd_iommu_print_event. Many of the changes
are related to stopping the abuse of %p to print hex values that are no
pointers in the hypervisor address space.

CC: Valentine Sinitsyn <valentine.sinitsyn@gmail.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>