core: ivshmem: Remove bogus MMIO region unregistrations
This is both wrong and unneeded:

ivshmem_disconnect_cell is called during cell destruction over a CPU of
the root cell. Therefore, the mmio_region_unregister will actually be
applied on the root cell, not the one that was owning the ivshmem
device. That is harmless in most cases if root cell and destroyed cell
have different memory layouts, thus do not map the ivshmem PCI resources
at the same address. If they do, and they did on AMD Seattle setup, we
remove a valid resource from the root cell and crash the latter on its
next access.

But the whole removal in ivshmem_disconnect_cell was actually never
required in first place because all MMIO regions of the cell under
destruction will simply vanish together with the same.

Fixes: 4b6f522c3218 ("core: ivshmem: Migrate to generic MMIO dispatcher")

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>