core: ivshmem: Clean up MSI-X capability definition
The lower 3 bits of the 2nd and 3rd DWORD of the MSI-X capability
contain the BAR Indicator Register. Those field encode the BAR number
used for MSI-X, and 4 is the one we use with ivshmem. There is no
relation to PCI_CFG_BAR as currently encoded, so clean things up. The
result will be the same value, so no functional change.

CC: Henning Schild <henning.schild@siemens.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>