core, tools: Add IRQ base number for virtual PCI hosts to cell config
In order to inject interrupts from virtual devices, we need to know the
(virtual) mapping of the VPCI host controller interrupts on the system
interrupt space. Introduce a per-cell config field that encodes this
base. It will only be used on targets without (suitable) physical MSI
support, which excludes x86.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>