core: ivshmem: Make device class tunable
This serves two purposes: First, there is an increasing difference
between the original IVSHMEM device (as implemented by QEMU) and our
version, still counting. Until we have come to a conclusion how to move
on with the device ("IVSHMEM 2.0" vs. something new), we need a way to
tell both devices apart, specifically in the Linux driver probing
process. That is one reason to switch to the PCI class "Others".

The other is that it gives us further 16 bits to encode which protocol
both peers are supposed to use over the device. We can now define this
via the cell configuration, informing both sides consistently about
which driver or application to be attached to a specific shared memory
device.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>