arm, arm64, x86: Add compile-time switch to crash cells on panic
This debugging feature is particularly useful with trying to get the
root cell running with a fresh config: If the root cell hits some
resources that the config generator didn't describe, you may want to
understand which part of the Linux accessed it. By simply crashing the
cell's CPU, you generally make Linux to dump a stack trace with the call
graph.

Crashing is achieved fairly reliably on all architectures by setting the
instruction pointer to 0. This only fails if the cell is running without
MMU, thus without NULL pointer detection.

CC: Valentine Sinitsyn <valentine.sinitsyn@gmail.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>