x86: Use parking page also for Intel
Consolidate the logic used for parking VCPUs to what is now used across
all architectures: park also Intel VCPUs by mapping in a cli+hlt loop
and letting guest execute it. This is what we already do on ARM.

Previously, we used the Intel-only feature of setting the VCPU into halt
state externally. In some rare cases, it was observed that this didn't
work as expected after a guest caused a violation and, thus, was parked.

CC: Valentine Sinitsyn <valentine.sinitsyn@gmail.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>