pci: avoid implicit device claiming with pci_add_dynid
Adding a dynid will cause probing for all matching devices not bound to
a driver yet. So one jailhouse_pci_claim_release(..CLAIM) had the
potential to claim all driver-less devices with the same IDs.

This patch keeps a list of devices explicitly claimed, all others will
be rejected in the probe-function.

Reported-by: Koehrer Mathias (ETAS/ESW5) <mathias.koehrer@etas.com>
Signed-off-by: Henning Schild <henning.schild@siemens.com>
[Jan: replace goto with break in jailhouse_pci_stub_probe]
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>