tools: Add ARM and ARM64 support to Linux loader script
This extends "jailhouse cell linux" with support for ARM and ARM64
targets. Those need a device tree template to describe the cell, so we
extend the command line options accordingly.

Command line and initrd location/size are passed to the kernel via the
device tree. Therefore, we have to parse the input dtb and patch/extend
it with the required properties under /chosen.

Furthermore, placing the images into the cell memory has arch-dependent
constraints. We need to leave some space (we chose 3-times the kernel
image) between kernel and initrd so that Linux can do decompression and
relocation. And on ARM64, the kernel has to be put on 2 MB-aligned
address plus it load offset.

As we are already patching the dtb, we can also create the /memory node
from the cell configuration. This avoids that config and device tree
have to be manually kept in sync in this regard. Just make sure to not
report any tiny region at the start of address space which we need for
the loader but Linux cannot use and rather get confused by, failing to
boot.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>