configs: Tune AMD Seattle Linux inmate configuration
Be less wasteful with memory: provide most of the RAM 1:1 mapped and
report that as available memory via the device tree. Also set
JAILHOUSE_MEM_DMA for the RAM regions, even though we don't have SMMU
support yet, so that the Linux loader script can identify them as RAM.

As jailhouse-cell-linux is now writing the /memory node, we can remove
it from the device tree, instead of updating it.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>