arm, arm64: Fix reset of non-root SMP cells
So far, we initialize cpu_on_entry during arch_cell_create, but not on
arch_cell_reset which breaks the PSCI state after reload and starting a
cell. We can simply move the related code over because arch_cell_create
is issued by cell_start, thus early enough after cell creation. The ARM
and ARM64 implementations of the initialization code look different but
already did the same.

And the bonus is even a few lines of code less.

Fixes: 58a0e22f08c8 ("arm: Introduce infrastructure for reworked CPU
                      management")

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>