configs: Add non-root Linux demo cells and device trees for Jetson TK1
We use the same UART for the non-root Linux as the root cell to simplify
the physical setup. Can be changed as needed.

To test SMP support, we hand out two CPUs. Note the peculiarity that the
first CPU has to have ID 0 while the second be reported with its real
one, otherwise Linux won't pick up both. Reason is still unclear.

On ARM, we have the need to align the virtual start address of the
identity-mapped memory to 128 MB. Furthermore, Linux won't boot if that
region starts at 0xf0000000 or higher. This constrains will be removed
at least from the Jetson as soon as we add SMMU support.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>