core: ivshmem: Fix matching of peers
So far, we created another ivshmem_data structure in case two devices
matched by BDF but not by their shared memory regions. This did not make
sense (ivshmem_find only returns the first BDF match) and it could leak
the data structure when the device was destroyed again.

Change the policy so that we match on BDF but fail creating the second
device in case of region mismatch.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>