inmates: Delay uart init if it is shared with the hypervisor
If we share an inmate UART with the hypervisor console, both may write
simultaneously during startup, causing mangled output. We used to
prevent this in the apic-demo until 211558763ddd by waiting for the
output to be idle for a couple of checks. Reintroduce this feature, but
now for all archs.

A missing divider is a good indicator for UART sharing.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>