arm-common: Mask and deactivate all SPIs on cell reset and destruction
Bring the GICD into a safe state when a cell is reset or destroyed: All
its SPIs have to be masked and also deactivated so that they do not fire
unexpectedly or are stuck in activation.

Deactivation is the architecturally defined reset state while masking
(disabling) is implementation defined - so we define it to be masked
under Jailhouse.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>