core: ivshmem: Adjust MSI-X region size calculation
The length of a PCI resource has be power of 2. Our MSI-X region size is
primarily defined by the table size where each vector requires 16. The
PBA region will always be smaller than that. So we can dimension the
region correctly by taking the MSI-X table twice.

This does not have an effect on the region size which stays at 0x20 with
a single MSI-X vector.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>