x86: Fix potentially inconsistent IOAPIC entries
When writing the upper half of a redirection entry while it is masked,
we do not update the physical register so far nor try to map the entry -
which could fail due to an inconsistent state. However, when we then
take the update for the lower half, the upper part will still not be
written. Fix this by writing the upper half along with the lower one.

Likely, we were lucky so far that only the root cell uses the IOAPIC,
and the upper half was left behind from the handover in a usable state.

Fixes: f651754c72e3 ("x86: Virtualize IOAPIC redir table for interrupt remapping support")
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>