Update config files for TX1
This patch updates the config files for the Nvidia TX1 board.
Besides enlarging the amount of memory for the root cell, the patch
fixes an issue with pulseaudio on X11 and error messages printed by
the memory controller driver on the stock Nvidia 3.10 kernel.
Tested on both the Nvidia 3.10 and the Vanilla 4.9 kernels.

Signed-off-by: Claudio Scordino <claudio@evidence.eu.com>
Signed-off-by: Errico Guidieri <e.guidieri@evidence.eu.com>
Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>