x86, inmates: Fix potential stack corruptions via red zones
By default, compilers for the x86-64 ABI can use a 128-byte zone below
the stack pointer as temporary memory. Anything that interrupts the
current context is supposed to respect that area, i.e. increment rsp
before using the stack.

As we are using no separate stacks for interrupts (including NMIs)
generated in the hypervisor as well as the inmates, we risk infrequent
corruptions like reported in

https://www.mail-archive.com/jailhouse-dev@googlegroups.com/msg02195.html

And we cannot adjust the stack pointer even right at the beginning of
the interrupt entry point because the processor already saved the basic
context for us. The best fix is to disable the red zone, just like Linux
does as well.

A quick scan of the code generated right now by gcc 4.8.5 for
jailhouse-intel.bin only revealed on usage of the red zone by
vmcs_clear, a function used during cell create and destruction. However,
mileage can vary in other configurations.

Signed-off-by: Jan Kiszka <jan.kiszka@siemens.com>